Refactor service.rs into smaller modules
